// types.ts
// Zod
import z from "zod";

// RHF
import { FieldPath, UseFormReturn } from "react-hook-form";

// Zod schemas
import { InvoiceSchema, ItemSchema } from "@/lib/schemas";

// Form types
export type InvoiceType = z.infer<typeof InvoiceSchema>;
export type ItemType = z.infer<typeof ItemSchema>;
export type FormType = UseFormReturn<InvoiceType>;
export type NameType = FieldPath<InvoiceType>;
export type CurrencyType = {
  [currencyCode: string]: string;
};

// Signature types
export type SignatureColor = {
  name: string;
  label: string;
  color: string;
};

export type SignatureFont = {
  name: string;
  variable: string;
};

export enum SignatureTabs {
  DRAW = "draw",
  TYPE = "type",
  UPLOAD = "upload",
}

// Wizard types
export type WizardStepType = {
  id: number;
  label: string;
  isValid?: boolean;
};

// Export types
export enum ExportTypes {
  KV = "KV",
  JSON = "JSON",
  CSV = "CSV",
  XML = "XML",
  XLSX = "XLSX",
  DOCX = "DOCX",
}